__unzipped/guardian/requirements.txt
python-telegram-bot==20.6
telethon==1.33.1
sqlalchemy==2.0.23
python-dotenv==1.0.1
pytest==7.4.3
pytest-asyncio==0.23.6
